refactor: (WIP) resolving issues with POST /invoices/search
