+++ app/controllers/invoices_controller.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class InvoicesController < ApplicationController
+  skip_before_action :verify_authenticity_token, only: %i[search]
+
   include MaybeAccountSpecific
 
   load_account :all,
