+++ app/frontend/features/InvoiceManager/api/getInvoices.ts
@@ -1,19 +1,8 @@
-import { Invoice } from '../types'
+import { InvoiceSearchProps } from '../types'
 import { composeQueryParams } from '../utils'
 
-type SortKey = keyof Pick<Invoice, 'account' | 'status' | 'dueAt' | 'paidAt' | 'amount'> | 'account.email' | 'account.displayName'
-type FilterKey = 'status' | 'invoiceNumber' | 'account.displayName' | 'account.email'
-
-export interface InvoiceSearchProps {
-  // Sorting
-  s: Partial<Record<SortKey, 'asc' | 'desc' | null>>
-  // Filtering
-  f: Partial<Record<FilterKey, string>>
-  // (Full-text search) query string
-  q: string
-}
-
 export const getInvoices = async (payload?: Partial<InvoiceSearchProps>) => {
+  console.debug(`Fetching invoices with payload:`, { ...payload })
   const params = composeQueryParams(payload ?? {}, new URLSearchParams({ format: 'json' }))
   const result = await fetch(`/invoices?${params.toString()}`)
   const data = await result.json()
