+++ app/frontend/features/InvoiceManager/hooks/useInvoiceSearchQuery.ts
@@ -1,5 +1,5 @@
-import { getInvoices, InvoiceSearchProps } from '@/utils/api'
-import { Invoice } from '../types'
+import { findInvoices } from '@/utils/api'
+import { Invoice, InvoiceSearchProps } from '../types'
 import { useQuery } from '@tanstack/react-query'
 import { useState } from 'react'
 
@@ -9,7 +9,7 @@ export const useInvoiceSearchQuery = (props?: Partial<InvoiceSearchProps>) => {
   const query = useQuery({
     queryKey: ['invoiceSearch'],
     queryFn: async () => {
-      const data = await getInvoices(props)
+      const data = await findInvoices(props)
       setInvoices(data)
       return data
     },
